SET the Hold request

+++ Project492/client/src/components/customer/page/decide.js
@@ -16,6 +16,9 @@ function Decide() {
     const yes = async(e) => {
         try {
             e.preventDefault()
+            const check  = window.confirm('Confirm ?!!')
+             if(!check) return ;
+
             await socket.emit('customer-select', {order_id : order.id ,select : true})
             await axios.put(`/order/choose/${order.id}`)
 
@@ -30,10 +33,14 @@ function Decide() {
         try {
             e.preventDefault()
 
+            const check  = window.confirm('Sure ?!!')
+            if(!check) return ;
+
             await socket.emit('customer-select', {order_id : order.id ,select : false})
             await axios.delete(`/order/${order.id}`)
             // navigate('/index')
-            return navigate('/') // problem is here !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+             // problem is here !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+            return navigate('/index')
             // window.location.reload()
